Conversion From BigQuery to PSQL (#1)
Converts the storage of the Peloton metrics data from Google BigQuery to a PostgreSQL database. This was done to enable this application to be run locally without a dependency on Google Cloud (or any cloud platform for that matter).

Anybody should be now able to clone this repo and have it up and running without the need to setup a cloud project.